+++ cucumber-tsflow/src/cucumber/binding-registry.ts
@@ -1,7 +1,8 @@
 import { ISupportCodeLibrary } from '@cucumber/cucumber/lib/support_code_library_builder/types';
 import _ from 'underscore';
 import { StepBinding, StepBindingFlags } from '../types/step-binding';
-import { ContextType, StepPattern, TagName } from '../types/types';
+import { ContextType } from '../types/context-injection';
+import { StepPattern, TagName } from '../types/types';
 
 /**
  * Describes the binding metadata that is associated with a binding class.